requirements.txt
fastapi
uvicorn
python-dotenv
aiofiles
azure-storage-blob
azure-ai-formrecognizer
python-multipart
openai
requests
azure-identity
azure-ai-projects
azure-ai-agents
bcrypt
pymongo
pymupdf
email-validator
python-docx
motor
reportlab
azure.communicaion.s